added feature to set input type to single/bulk
I removed the input_file cmd line argument and instead, as you suggested, added -it argument where the user can specify single/bulk.
If single is given, then the url will be the next argument. Else if bulk is given as input type then the .txt file with the urls is given as the next argument.